gcf_deploy/requirements.txt
# Google Cloud Functions Requirements
# Deploy with: gcloud functions deploy send_daily_quote --requirements-file=gcf_requirements.txt

# Core dependencies
python-telegram-bot==21.0
anthropic>=0.40.0
apscheduler==3.10.4
sqlalchemy>=2.0.0
python-dotenv==1.0.0

# Google Cloud Functions dependencies
flask>=3.0.0
